refactor(NodeService): create NodeInfoService and remove UpgradeModule #615
