+++ src/assets/wise5/components/conceptMap/concept-map-authoring/concept-map-authoring.module.ts
@@ -11,6 +11,7 @@ import { EditComponentPrompt } from '../../../../../app/authoring-tool/edit-comp
 import { ProjectAssetService } from '../../../../../app/services/projectAssetService';
 import { AnnotationService } from '../../../services/annotationService';
 import { ConfigService } from '../../../services/configService';
+import { NodeInfoService } from '../../../services/nodeInfoService';
 import { NodeService } from '../../../services/nodeService';
 import { ProjectService } from '../../../services/projectService';
 import { SessionService } from '../../../services/sessionService';
@@ -39,6 +40,7 @@ import { ConceptMapAuthoring } from './concept-map-authoring.component';
     ConceptMapService,
     ConfigService,
     NodeService,
+    NodeInfoService,
     ProjectAssetService,
     ProjectService,
     SessionService,
