+++ src/assets/wise5/components/draw/draw-authoring/draw-authoring.module.ts
@@ -11,6 +11,7 @@ import { EditComponentPrompt } from '../../../../../app/authoring-tool/edit-comp
 import { ProjectAssetService } from '../../../../../app/services/projectAssetService';
 import { AnnotationService } from '../../../services/annotationService';
 import { ConfigService } from '../../../services/configService';
+import { NodeInfoService } from '../../../services/nodeInfoService';
 import { NodeService } from '../../../services/nodeService';
 import { ProjectService } from '../../../services/projectService';
 import { SessionService } from '../../../services/sessionService';
@@ -39,6 +40,7 @@ import { DrawAuthoring } from './draw-authoring.component';
     ConfigService,
     DrawService,
     NodeService,
+    NodeInfoService,
     ProjectAssetService,
     ProjectService,
     SessionService,
