+++ src/assets/wise5/components/showGroupWork/show-group-work-authoring/show-group-work-authoring.component.spec.ts
@@ -18,6 +18,7 @@ import { UtilService } from '../../../services/utilService';
 import { ShowGroupWorkAuthoringComponent } from './show-group-work-authoring.component';
 import { EditComponentPrompt } from '../../../../../app/authoring-tool/edit-component-prompt/edit-component-prompt.component';
 import { MatCheckboxModule } from '@angular/material/checkbox';
+import { NodeInfoService } from '../../../services/nodeInfoService';
 
 describe('ShowGroupWorkAuthoringComponent', () => {
   let component: ShowGroupWorkAuthoringComponent;
@@ -42,6 +43,7 @@ describe('ShowGroupWorkAuthoringComponent', () => {
       providers: [
         ConfigService,
         NodeService,
+        NodeInfoService,
         ProjectAssetService,
         ProjectService,
         SessionService,
