+++ src/assets/wise5/components/showMyWork/show-my-work-authoring/show-my-work-authoring.component.spec.ts
@@ -11,6 +11,7 @@ import { UpgradeModule } from '@angular/upgrade/static';
 import { EditComponentPrompt } from '../../../../../app/authoring-tool/edit-component-prompt/edit-component-prompt.component';
 import { ProjectAssetService } from '../../../../../app/services/projectAssetService';
 import { ConfigService } from '../../../services/configService';
+import { NodeInfoService } from '../../../services/nodeInfoService';
 import { NodeService } from '../../../services/nodeService';
 import { ProjectService } from '../../../services/projectService';
 import { SessionService } from '../../../services/sessionService';
@@ -51,6 +52,7 @@ describe('ShowMyWorkAuthoringComponent', () => {
       providers: [
         ConfigService,
         NodeService,
+        NodeInfoService,
         ProjectAssetService,
         ProjectService,
         SessionService,
