+++ src/assets/wise5/vle/vle.component.spec.ts
@@ -36,6 +36,7 @@ import { NodeStatusIcon } from '../themes/default/themeComponents/nodeStatusIcon
 import { NodeIconComponent } from '../classroomMonitor/classroomMonitorComponents/shared/nodeIcon/node-icon.component';
 import { Node } from '../common/Node';
 import { FormsModule } from '@angular/forms';
+import { NodeInfoService } from '../services/nodeInfoService';
 
 let component: VLEComponent;
 let fixture: ComponentFixture<VLEComponent>;
@@ -86,6 +87,7 @@ describe('VLEComponent', () => {
         ConfigService,
         ComponentService,
         NodeService,
+        NodeInfoService,
         NotebookService,
         NotificationService,
         ProjectService,
