+++ src/messages.xlf
@@ -15981,8 +15981,8 @@ If this problem continues, let your teacher know and move on to the next activit
       <trans-unit id="2821363608497156596" datatype="html">
         <source>Step Info</source>
         <context-group purpose="location">
-          <context context-type="sourcefile">src/assets/wise5/services/nodeService.ts</context>
-          <context context-type="linenumber">723</context>
+          <context context-type="sourcefile">src/assets/wise5/services/nodeInfoService.ts</context>
+          <context context-type="linenumber">16</context>
         </context-group>
       </trans-unit>
       <trans-unit id="1384916903989385807" datatype="html">
